updates for MuTau FF closure and application
