+++ calculate_functions.py
@@ -178,3 +178,8 @@ def highest_mjj_pair(TLorentzVector_Jets):
         j2_idx = k_jet
   if (j1_idx*j2_idx < 0): print("jet index unassigned!")
   return TLorentzVector_Jets[j1_idx], TLorentzVector_Jets[j2_idx]
+
+def user_exp(x, a, b, c, d):
+    return a*np.exp(-b*(x-c)) + d
+
+
