+++ cut_and_study_functions.py
@@ -9,6 +9,7 @@ from utility_functions   import time_print, text_options, log_print
 from cut_ditau_functions import make_ditau_cut 
 from cut_mutau_functions import make_mutau_cut, make_mutau_TnP_cut
 from FF_functions        import make_ditau_SR_cut, make_ditau_AR_cut, make_mutau_SR_cut, make_mutau_AR_cut
+from FF_functions        import set_FF_values, add_FF_weights
 from cut_etau_functions  import make_etau_cut,  make_etau_AR_cut
 from cut_dimuon_functions import make_dimuon_cut
 
@@ -174,186 +175,6 @@ def append_flavor_indices(event_dictionary, final_state_mode, keep_fakes=False):
   return event_dictionary
 
 
-def set_FF_values(final_state_mode, jet_mode_and_DeepTau_version, determining_FF):
-  '''
-  '''
-  # should have aiso/iso as well
-  FF_values = {
-    # FS : { "jet_mode" : [intercept, slope] }  
-    "ditau" : { 
-      # updated March 19th
-      "custom_0j_2p5_FF"    :    [0.278,-0.000577],
-      "custom_1j_2p5_FF"    :    [0.230,-0.000493],
-      "custom_GTE2j_2p5_FF" :    [0.236,-0.000953],
-      ### from V12 samples
-      #"custom_0j_2p5_FF" :    [0.278,-0.000577],
-      #"custom_1j_2p5_FF" :    [0.231,-0.000509],
-      #"custom_GTE2j_2p5_FF" :    [0.237,-0.000970],
-      #"custom_0j_2p5_FF" :    [0.277605,-0.000507954],
-      #"custom_1j_2p5_FF" :    [0.245381,-0.00054067],
-      #"custom_GTE2j_2p5_FF" : [0.232493,-0.000626524],
-
-      "custom_0j_2p5_CH" :    [1.03281, 0.000366355], # i think this is an underestimate
-      "custom_1j_2p5_CH" :    [1.08724, 3.92753e-05], # same here
-      "custom_GTE2j_2p5_CH" : [1.11779, 0.000141817],
-      ###
-    },
-    "mutau" : {  # wrong 2p5
-      "0j_2p5"     : [0.037884, 0.000648851],
-      "1j_2p5"     : [0.0348384, 0.000630731],
-      "GTE2j_2p5"  : [0.0342287, 0.000358899],
-
-      #"custom_0j_2p5_FF" :    [0.0262364,-1.88738e-05], # combined fit, by hand
-      #"custom_1j_2p5_FF" :    [0.0262364,-1.88738e-05],
-      #"custom_GTE2j_2p5_FF" : [0.0262364,-1.88738e-05],
-
-      #"custom_0j_2p5_FF" :    [0.0790,0.000444], # combined fit, w framework, QCD only
-      #"custom_1j_2p5_FF" :    [0.0790,0.000444],
-      #"custom_GTE2j_2p5_FF" : [0.0790,0.000444],
-
-      "custom_0j_2p5_FF"    : [0.0827,0.000420], #split by jet, w framework, QCD only
-      "custom_1j_2p5_FF"    : [0.110,-0.000591], 
-      "custom_GTE2j_2p5_FF" : [0.108,0], # set slope to zero otherwise negative QCD... 
-
-      "custom_0j_2p5_CH"    : [1.1, 0], # ad-hoc scaling
-      "custom_1j_2p5_CH"    : [1.1, 0],
-      "custom_GTE2j_2p5_CH" : [1.1, 0],
-
-    },
-    "etau"  : {#Dummy values
-      "0j_2p5"     : [1, 1], 
-      "1j_2p5"     : [1, 1],
-      "GTE2j_2p5"  : [1, 1],
-
-      "custom_0j_2p5_FF" : [0.0483552, -4.68741e-05], # combined fit, by hand
-      "custom_1j_2p5_FF" : [0.0483552, -4.68741e-05],
-      "custom_GTE2j_2p5_FF" : [0.0483552, -4.68741e-05],
-
-      "custom_0j_2p5_CH"    : [1.1, 0], # ad-hoc scaling
-      "custom_1j_2p5_CH"    : [1.1, 0],
-      "custom_GTE2j_2p5_CH" : [1.1, 0],
-
-    },
-  } 
-  if (determining_FF == True):
-    print("determining FF dictionary, setting dummy values and not plotting QCD")
-    intercept, slope = 0.2, 1.1
-  else:
-    intercept = FF_values[final_state_mode][jet_mode_and_DeepTau_version][0]
-    slope     = FF_values[final_state_mode][jet_mode_and_DeepTau_version][1]
-    #slope2    = FF_values[final_state_mode][jet_mode_and_DeepTau_version][2]
-
-  #return intercept, slope, slope2
-  return intercept, slope
-
-
-def add_FF_weights(event_dictionary, final_state_mode, jet_mode, DeepTau_version, determining_FF=False, bypass=None):
-  '''
-  '''
-  unpack_FFVars = ["Lepton_pt", "HTT_m_vis", "l1_indices", "l2_indices"]
-  unpack_FFVars = (event_dictionary.get(key) for key in unpack_FFVars)
-  to_check = [range(len(event_dictionary["Lepton_pt"])), *unpack_FFVars]
-  FF_weights = []
-  bins = [50, 60, 70, 80, 90, 100, 110, 120, 130, 140, 150, 160, 170, 180, 190, 200, 250, 300]
-  ditau_weight_map = {
-    # these are all very close to 0.999, which means in the most pessimistic case
-    # you lose 10 events in every 10000 events.
-    # since we have at most 90000 events in the 0j case, the biggest error we can have is
-    # 90 events. This is not big enough to effect plots dramatically, so as long as the
-    # probs are close to 0.999, they don't need to be updated every time
-    # SHOULD update them the final time
-    "Inclusive_2p5" : [bins, # estimated, values not derived
-                 [0.999, # < 50
-                  0.999, 0.999, 0.999, 0.999, 0.999, 0.999, 0.999, 0.999,
-                  0.999, 0.999, 0.999, 0.999, 0.999, 0.999, 0.999,
-                  0.999, 0.999, 0.999]], # > 200
-    "0j_2p5" : [bins,
-          [0.998869, # < 50
-           0.999233, 0.999735, 0.999866, 0.999873, 0.999901, 0.999944, 0.999944, 0.999939, 
-           0.999930, 0.999924, 0.999921, 0.999913, 0.999900, 0.999899, 0.999885, 
-           0.999871, 0.999826, 0.999691]], # > 200
-    "1j_2p5" : [bins,
-          [0.999676, 
-           0.999512, 0.999607, 0.999771, 0.999810, 0.999827, 0.999864, 0.999851, 0.999826, 
-           0.999807, 0.999795, 0.999776, 0.999775, 0.999743, 0.999740, 0.999726, 
-           0.999700, 0.999630, 0.999492]],
-    "GTE2j_2p5" : [bins,
-          [0.999411, 
-           0.999046, 0.999038, 0.999246, 0.999336, 0.999382, 0.999425, 0.999362, 0.999296, 
-           0.999252, 0.999181, 0.999143, 0.999101, 0.999101, 0.999038, 0.999003, 
-           0.998991, 0.998904, 0.998680]],
-    "0j_2p1" : [bins,
-          [0.996032, 
-           0.997514, 0.999023, 0.999554, 0.999771, 0.999772, 0.999850, 0.999858, 
-           0.999850, 0.999843, 0.999824, 0.999811, 0.999792, 0.999783, 0.999781, 0.999735, 
-           0.999683, 0.999594, 0.999428]],
-    "1j_2p1" : [bins,
-          [0.998659, 
-           0.998212, 0.998626, 0.999254, 0.999635, 0.999633, 0.999658, 0.999630, 
-           0.999600, 0.999549, 0.999526, 0.999469, 0.999428, 0.999394, 0.999348, 0.999345, 
-           0.999223, 0.999137, 0.998901]],
-    "GTE2j_2p1" : [bins,
-          [0.997985, 
-           0.997208, 0.997390, 0.998144, 0.998912, 0.998832, 0.998663, 0.998491, 
-           0.998358, 0.998206, 0.998065, 0.998010, 0.997843, 0.997769, 0.997668, 0.997625, 
-           0.997477, 0.997151, 0.997048]],
-  }
-  
-  FF_key = jet_mode + "_" + DeepTau_version
-
-  if (bypass != None):
-    intercept = bypass[0]
-    slope     = bypass[1]
-    OSSS_bias_intercept = bypass[2]
-    OSSS_bias_slope     = bypass[3]
-    ditau_weight_map[FF_key] = {}
-    ditau_weight_map[FF_key] = [bins,[1 for i in range(18)]]
-
-  else:
-    intercept, slope = set_FF_values(final_state_mode, "custom_"+jet_mode+"_2p5_FF", determining_FF)
-    OSSS_bias_intercept, OSSS_bias_slope = set_FF_values(final_state_mode, "custom_"+jet_mode+"_2p5_CH", determining_FF)
-
-  #intercept, slope = set_FF_values("ditau", FF_key)
-  #closure_intercept, closure_slope = set_FF_values("ditau", "Closure_" + DeepTau_version)
-  #OSSS_bias_intercept, OSSS_bias_slope = set_FF_values("ditau", "OSSS_Bias_" + DeepTau_version)
-
-  #intercept, slope, slope2 = set_FF_values("ditau", "custom_"+jet_mode+"_2p5_check_FF")
-  #closure_intercept, closure_slope, closure_slope2 = set_FF_values("ditau", "custom_"+jet_mode+"_2p5_check_Clos")
-  #OSSS_bias_intercept, OSSS_bias_slope, OSSS_bias_slope2 = set_FF_values("ditau", "custom_"+jet_mode+"_2p5_check_CH")
-
-  for i, lep_pt, m_vis, l1_idx, l2_idx in zip(*to_check):
-    if m_vis < bins[0]: # 50
-      one_minus_MC_over_data_weight = ditau_weight_map[FF_key][1][0] # first weight
-    elif m_vis > bins[-3]: # > 200
-      if m_vis > bins[-1]: # > 300
-        one_minus_MC_over_data_weight = ditau_weight_map[FF_key][1][-1] # last weight
-      elif bins[-2] < m_vis < bins[-1]: # between 250 and 300
-        one_minus_MC_over_data_weight = ditau_weight_map[FF_key][1][-2]
-      elif bins[-3] < m_vis < bins[-2]: # between 200 and 250
-        one_minus_MC_over_data_weight = ditau_weight_map[FF_key][1][-3]
-    else: # mvis between 50 and 200
-      m_vis_idx = int(m_vis // 10) - 5 # makes 50 bin zero idx
-      m_vis_weight_idx = m_vis_idx + 1 # 0 in weights is < 50 weight
-      one_minus_MC_over_data_weight = ditau_weight_map[FF_key][1][m_vis_weight_idx]
-
-    #l1_pt = lep_pt[l1_idx] if lep_pt[l1_idx] < 120.0 else 120.0
-    #l2_pt = lep_pt[l2_idx] if lep_pt[l2_idx] < 200.0 else 200.0
-    l1_pt = lep_pt[l1_idx]
-    l2_pt = lep_pt[l2_idx]
-    m_vis = m_vis if m_vis < 350.0 else 350.0
-    FF_weight = one_minus_MC_over_data_weight*(intercept + l1_pt * slope)
-    #FF_weight *= (closure_intercept + lep_pt[l2_idx] * closure_slope)
-    FF_weight *= (OSSS_bias_intercept + m_vis * OSSS_bias_slope)
-
-    #FF_weight = one_minus_MC_over_data_weight*(intercept + l1_pt * slope + l1_pt*l1_pt*slope2)
-    #FF_weight *= (closure_intercept + l2_pt * closure_slope + l2_pt*l2_pt*closure_slope2)
-    #FF_weight *= (OSSS_bias_intercept + m_vis * OSSS_bias_slope + m_vis*m_vis*OSSS_bias_slope2)
-
-    FF_weights.append(FF_weight)
-  event_dictionary["FF_weight"] = np.array(FF_weights)
-  return event_dictionary
-
-
 def make_jet_cut(event_dictionary, jet_mode):
   nEvents_precut = len(event_dictionary["Lepton_pt"])
   unpack_jetVars = ["nCleanJet", "CleanJet_pt", "CleanJet_eta", "CleanJet_phi", "CleanJet_mass"]
@@ -490,8 +311,8 @@ def make_jet_cut(event_dictionary, jet_mode):
     # importantly different from inclusive
     #event_dictionary["pass_2j_cuts"]    = np.array(pass_2j_cuts)
     #event_dictionary["pass_3j_cuts"]    = np.array(pass_3j_cuts)
-    event_dictionary["pass_GTE1j_cuts"]    = np.array(pass_GTE1j_cuts)
-    #event_dictionary["pass_GTE2j_cuts"]    = np.array(pass_GTE2j_cuts)
+    #event_dictionary["pass_GTE1j_cuts"]    = np.array(pass_GTE1j_cuts)
+    event_dictionary["pass_GTE2j_cuts"]    = np.array(pass_GTE2j_cuts)
     event_dictionary["CleanJetGT30_pt_1"]  = np.array(CleanJetGT30_pt_1)
     event_dictionary["CleanJetGT30_pt_2"]  = np.array(CleanJetGT30_pt_2)
     #event_dictionary["CleanJetGT30_pt_3"]  = np.array(CleanJetGT30_pt_3)
@@ -505,9 +326,9 @@ def make_jet_cut(event_dictionary, jet_mode):
     event_dictionary["FS_detajj"] = np.array(detajj_array)
 
   # can only do this if inclusive
-  if jet_mode == "Inclusive":
-    print("nEvents with exactly 0 and ≥1 jets")
-    print(f"{len(np.array(pass_0j_cuts))}, {len(np.array(pass_GTE1j_cuts))}")
+  #if jet_mode == "Inclusive":
+    #print("nEvents with exactly 0 and ≥1 jets")
+    #print(f"{len(np.array(pass_0j_cuts))}, {len(np.array(pass_GTE1j_cuts))}")
 
   return event_dictionary
 
@@ -756,9 +577,12 @@ def apply_AR_cut(process, event_dictionary, final_state_mode, jet_mode, DeepTau_
     protected_branches = set_protected_branches(final_state_mode=final_state_mode, jet_mode="none")
     event_dictionary   = apply_cut(event_dictionary, "pass_cuts", protected_branches)
     #
-    event_dictionary   = add_FF_weights(event_dictionary, final_state_mode, jet_mode, DeepTau_version,
-                                        determining_FF=determining_FF)
-    determining_FF = False # paranoid thing to prevent leaking the variable
+    semilep_mode = "QCD"
+    event_dictionary   = add_FF_weights(event_dictionary, final_state_mode, jet_mode, semilep_mode, full_FF=False)
+    # old version
+    #event_dictionary   = add_FF_weights(event_dictionary, final_state_mode, jet_mode, DeepTau_version,
+    #                                    determining_FF=determining_FF)
+    #determining_FF = False # paranoid thing to prevent leaking the variable
   else:
     print(f"{final_state_mode} : {jet_mode} not possible. Continuing without AR or FF method applied.")
   return event_dictionary
